backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["TobiFank <tobias7402@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]  # Add this line

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
uvicorn = "^0.32.0"
alembic = "^1.14.0"
httpx = "^0.27.2"
tenacity = "^9.0.0"
pydantic-settings = "^2.6.1"
pymilvus = "^2.4.9"
numpy = "^2.1.3"
pillow = "^11.0.0"
pypdf = "^5.1.0"
pytesseract = "^0.3.13"
python-docx = "^1.1.2"
python-magic = "^0.4.27"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
